created function
Began creating the function that would search the data. It now looks for the flight being searched for, and will check if the user has entered data into the text field. Next I need to grab the data from the flight and display it to the page, and I also need to handle what happens when a flight can't be found.