display flight function
Moved the HTML that holds the flight info into its own function to make the findFlight function cleaner. Also handle a status that doesn't return 200r in getJSOnData - maybe that can be better I will look at it later.

TODO: Slap a nice coat of paint on it to make it look nicer than generic BS.